requirements.txt
boto3==1.9.235
botocore==1.12.235
Collectfast==1.0.0
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.1.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages==1.7.2
django-summernote==0.8.11.4
django-tinymce==2.8.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
drf-nested-routers==0.91
entrypoints==0.3
flake8==3.7.8
gunicorn==19.9.0
jmespath==0.9.4
mccabe==0.6.1
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
typing==3.7.4.1
typing-extensions==3.7.4
urllib3==1.25.6
whitenoise==4.1.4
